Organize components into their own folders
