+++ pages/register.js
@@ -5,7 +5,7 @@ import { useRouter } from "next/router";
 import { v4 as uuidv4 } from "uuid";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import Layout from "../components/layout";
+import Layout from "../components/Layout/Layout";
 import { UserContext } from "../contexts/UserContext";
 
 export default function Register() {
