+++ pages/topics/[slug]/[id].js
@@ -1,21 +1,23 @@
 import { useState, useEffect } from "react";
 import Head from "next/head";
 import Link from "next/link";
-import Layout from "../../../components/layout";
-import Post from "../../../components/post";
-import PostForm from "../../../components/postForm";
-import PostEditForm from "../../../components/postEditForm";
+import Layout from "../../../components/Layout/Layout";
+import Post from "../../../components/Post/Post";
+import PostForm from "../../../components/PostForm/PostForm";
+import PostEditForm from "../../../components/PostEditForm/PostEditForm";
 
 export default function SingleTopic({ topic, posts }) {
   const { id, title, slug, category_name, category_id, username } = topic;
+
   const [isPostFormOpen, setIsPostFormOpen] = useState(false);
   const [isEditFormOpen, setIsEditFormOpen] = useState(false);
   const [currentPosts, setCurrentPosts] = useState(posts);
+  const [postToEdit, setPostToEdit] = useState({});
+
   const openPostForm = () => setIsPostFormOpen(true);
   const openEditForm = () => setIsEditFormOpen(true);
   const closePostForm = () => setIsPostFormOpen(false);
   const closeEditForm = () => setIsEditFormOpen(false);
-  const [postToEdit, setPostToEdit] = useState({});
 
   return (
     <div>
